chore: fix package declaration
