+++ extensions/common/auth/auth-authorization-oauth2-lib/src/main/java/org/eclipse/edc/api/authorization/filter/ScopeBasedAccessFilter.java
@@ -12,7 +12,7 @@
  *
  */
 
-package org.eclipse.edc.identityhub.api.authorization.filter;
+package org.eclipse.edc.api.authorization.filter;
 
 import jakarta.annotation.Priority;
 import jakarta.ws.rs.Priorities;
