+++ extensions/common/auth/auth-authorization-oauth2-lib/src/test/java/org/eclipse/edc/api/authorization/service/AuthorizationServiceImplTest.java
@@ -12,7 +12,7 @@
  *
  */
 
-package org.eclipse.edc.identityhub.api.authorization.service;
+package org.eclipse.edc.api.authorization.service;
 
 import jakarta.ws.rs.core.SecurityContext;
 import org.eclipse.edc.participantcontext.spi.types.AbstractParticipantResource;
